Add require_matching_rdns_lax check
A connecting host's EHLO domain sometimes doesn't match DRNS for the connecting IP; for example, the connection may come from mx21.example.org which may identify itself with EHLO mx.example.org. If mx.example.org resolves to a set of IPs including the connecting IP, this shouldn't be cause for concern, but it causes requireMatchingRDNS to fail since mx.example.org != mx21.example.org.

This check verifies that the EHLO domain resolves to the connecting IP, then separately looks up the PTR record for the connecting IP and verifies that the PTR record domain also resolves to the connecting IP. This blocks most spam coming from relays on compromised hosts.

Ideally this would just be an option to requireMatchingRDNS, but Maddy currently doesn't seem to support options for stateless checks.