Safelisting a message by its sender address/domain is only effective if we verify that the client IP address is authorized to send from that domain. Otherwise, anyone can spoof messages from that domain and bypass all other spam checks.
This patch changes safelisting so that all checks still run for each message (ensuring that any relevant headers from each check are still added to the message) but the results of each check are simply ignored. A safelisting check can now set RequireAuthResults=true to ensure that the message is still rejected/quarantined if it fails SPF/DKIM/DMARK authentication.