Multiple improvements to geobl check.
Removed blocklist_countries option; can now use one of `block_countries` (to specify a list of countries to disallow, allowing all others) or `allow_countries` (to specify a list of countries to allow, disallowing all others).

Also replaced `fail_open` with an `error_action` option to be more consistent with the rest of Maddy.